code/requirements.txt
Flask-SQLAlchemy
Flask-RESTful
Flask-JWT-Extended